Update travis infra.
Also tweak Travis build matrix.

This has no real effect except clarifying the display of Python-versions on
the Travis-CI status page.

This is pointless if you already read the tox output and see it's checking
against both versions of Python. But many people will see the shiny
Python-version labels on the Travis status page, and stop there - so
we should try to get those right, too.