Bugfix: Fix regression, failed parsing of ISO formatted timestamps.
'pull_isotime' was suddenly failing. iso8601 is very robust, switching
to astropy Time parser caused failures on encountering timestamp-strings
of with a '+0' timezone-signifier suffix.

I've also fixed the output so that we do not add the suffix when writing
new VOEvents, instead we convert to UTC then drop the timezone before
formatting.